requirements.txt
playwright==1.51.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.1.1
google-api-python-client==2.166.0
celery==5.4.0
redis==3.5.3
beautifulsoup4==4.12.3
loguru==0.7.3
python-dotenv==1.0.0